Samplers

Utility Samplers

Exact Sampler

QUBODrivers.ExactSampler.Optimizer — Type

ExactSampler.Optimizer{T}

This sampler performs an exhaustive search over all $2^{n}$ possible states.

Warn

Due to the exponetially large amount of visited states, it is not possible to use this sampler for problems any larger than $20$ variables big.

Random Sampler

QUBODrivers.RandomSampler.Optimizer — Type

RandomSampler.Optimizer{T}

Attributes

RandomSeed, "seed": Random seed to initialize the random number generator.

NumberOfReads, "num_reads": Number of random states sampled per run.

RandomGenerator, "rng": Random Number Generator instance.

Identity Sampler

QUBODrivers.IdentitySampler.Optimizer — Type

IdentitySampler.Optimizer{T}

This sampler selects precisely the state vector provided as warm-start.

Showcase

Before explaining in detail how to use this package, it's good to list a few examples for the reader to grasp. Below, there are links to the files containing the actual interface implementations. These are mostly thin wrappers interfacing with common algorithms and heuristics written in Python, Julia or C/C++.

Project	Source Code

DWave.jl	DWave

DWaveNeal.jl	DWaveNeal

IsingSolvers.jl	GreedyDescent

ILP

MCMCRandom

QuantumAnnealingInterface.jl	QuantumAnnealingInterface

CIMOptimizer.jl	CIMOptimizer